@@ -28,23 +28,23 @@
     "test:e2e": "playwright test"
   },
   "peerDependencies": {
-    "svelte": "^4.0.0"
+    "svelte": "^5.0.0"
   },
   "devDependencies": {
-    "@jill64/eslint-config-svelte": "1.3.10",
+    "@jill64/eslint-config-svelte": "2.0.1",
     "@jill64/npm-demo-layout": "1.0.249",
-    "@jill64/universal-sanitizer": "1.3.1",
-    "@jill64/playwright-config": "2.4.1",
+    "@jill64/universal-sanitizer": "1.3.5",
+    "@jill64/playwright-config": "2.4.2",
     "@jill64/prettier-config": "1.0.0",
-    "@jill64/sentry-sveltekit-cloudflare": "1.7.16",
-    "@playwright/test": "1.46.1",
-    "@sveltejs/adapter-cloudflare": "4.7.2",
-    "@sveltejs/kit": "2.5.24",
-    "svelte": "4.2.19",
-    "typescript": "5.5.4",
-    "vite": "5.4.2",
+    "@jill64/sentry-sveltekit-cloudflare": "2.0.1",
+    "@playwright/test": "1.49.1",
+    "@sveltejs/adapter-cloudflare": "5.0.0",
+    "@sveltejs/kit": "2.15.1",
+    "svelte": "5.16.0",
+    "typescript": "5.7.2",
+    "vite": "6.0.6",
     "vitest": "2.0.5",
-    "@sveltejs/vite-plugin-svelte": "3.1.2"
+    "@sveltejs/vite-plugin-svelte": "5.0.3"
   },
   "homepage": "",
   "author": {
