@@ -2,8 +2,10 @@
   import { Layout } from '@jill64/npm-demo-layout'
   import README from '../../README.md?raw'
   import packageJson from '../../package.json'
+
+  let { children } = $props()
 </script>
 
 <Layout {packageJson} {README}>
-  <slot />
+  {@render children()}
 </Layout>
